Fix sdl and linux build

+++ src/charset.cpp
@@ -144,7 +144,7 @@ Charset::printChar(uint32 c, int32 x, int32 y)
 	RWDEVICE::Im2DVertex *vert;
 	uint16 *ix;
 
-	if(c >= this->desc.count)
+	if((int32)c >= this->desc.count)
 		return;
 
 	if(this->raster != lastRaster || numChars >= NUMCHARS)
